requirements.txt
langchain==0.1.0
langchain-community==0.0.10
pypdf==3.17.4
pymupdf==1.23.8
python-docx==1.1.0